change return type on GetNumPending; JIRA DD-4200
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102299 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/structure/cd_utils/cuCdCore.hpp
@@ -198,7 +198,7 @@ public:
     int    GetRowsWithConsensus(vector<int>& consensusRows) const;
     bool   FindConsensusInSequenceList(vector<int>* indices = NULL) const;
 
-    int    GetNumPending() const {return(GetPending().size());}
+    size_t    GetNumPending() const {return(GetPending().size());}
 
 	//add aligns or sequences to CD
